requirements.txt
python-binance
binance-connector
websocket-client